Refactor the install/uninstall commands and helper functions to separate files to keep main cleaner.
