+++ cmd/certkit-agent/main_linux.go
@@ -3,24 +3,19 @@
 package main
 
 import (
-	"errors"
 	"flag"
 	"fmt"
 	"log"
 	"os"
-	"os/exec"
 	"os/signal"
-	"path/filepath"
-	"strings"
 	"syscall"
 
-	"github.com/certkit-io/certkit-agent/config"
-	"github.com/certkit-io/certkit-agent/utils"
+	agentinstall "github.com/certkit-io/certkit-agent/install"
 )
 
 const (
-	defaultUnitPath   = "/etc/systemd/system"
-	defaultConfigPath = "/etc/certkit-agent/config.json"
+	defaultUnitPath   = agentinstall.DefaultLinuxUnitPath
+	defaultConfigPath = agentinstall.DefaultLinuxConfigPath
 )
 
 func usageAndExit() {
@@ -41,79 +36,11 @@ Examples:
 }
 
 func installCmd(args []string) {
-	fs := flag.NewFlagSet("install", flag.ExitOnError)
-	serviceName := fs.String("service-name", defaultServiceName, "systemd service name")
-	unitDir := fs.String("unit-dir", defaultUnitPath, "systemd unit directory")
-	binPath := fs.String("bin-path", "", "path to certkit-agent binary (default: current executable)")
-	configPath := fs.String("config", defaultConfigPath, "path to config.json")
-	fs.Parse(args)
+	agentinstall.InstallLinux(args, defaultServiceName)
+}
 
-	mustBeRoot()
-
-	// Determine binary path (the installed binary path you want systemd to execute).
-	exe := *binPath
-	if exe == "" {
-		var err error
-		exe, err = os.Executable()
-		if err != nil {
-			log.Fatalf("failed to determine executable path: %v", err)
-		}
-		exe, err = filepath.EvalSymlinks(exe)
-		if err != nil {
-			log.Fatalf("failed to resolve executable symlinks: %v", err)
-		}
-	}
-
-	// Basic sanity checks.
-	if _, err := os.Stat(exe); err != nil {
-		log.Fatalf("binary path does not exist: %s (%v)", exe, err)
-	}
-	if !strings.HasPrefix(*unitDir, "/") {
-		log.Fatalf("--unit-dir must be an absolute path: %s", *unitDir)
-	}
-	if !strings.HasPrefix(*configPath, "/") {
-		log.Fatalf("--config must be an absolute path: %s", *configPath)
-	}
-
-	// Ensure config directory exists (config file contents are handled by your installer script).
-	if err := os.MkdirAll(filepath.Dir(*configPath), 0o755); err != nil {
-		log.Fatalf("failed to create config dir: %v", err)
-	}
-
-	// Ensure config exists or create it
-	if _, err := os.Stat(*configPath); os.IsNotExist(err) {
-		log.Printf("Config not found, creating %s", *configPath)
-		if err := config.CreateInitialConfig(*configPath); err != nil {
-			log.Fatalf("failed to create config: %v", err)
-		}
-	} else {
-		log.Printf("Config already exists at %s", *configPath)
-	}
-
-	if _, err := exec.LookPath("systemctl"); err != nil {
-		log.Printf("systemd not detected; skipping unit install. Run: %s run --config %s", exe, *configPath)
-		return
-	}
-
-	unitPath := filepath.Join(*unitDir, *serviceName+".service")
-	unitContent := renderSystemdUnit(exe, *configPath)
-
-	// Write unit file atomically.
-	if err := utils.WriteFileAtomic(unitPath, []byte(unitContent), 0o644); err != nil {
-		log.Fatalf("failed to write unit file %s: %v", unitPath, err)
-	}
-
-	// systemd: daemon-reload, enable, start
-	if err := runCmdLogged("systemctl", "daemon-reload"); err != nil {
-		log.Fatalf("systemctl daemon-reload failed: %v", err)
-	}
-	if err := runCmdLogged("systemctl", "enable", "--now", *serviceName+".service"); err != nil {
-		log.Fatalf("systemctl enable --now failed: %v", err)
-	}
-
-	machineId, _ := utils.GetStableMachineID()
-	log.Printf("Installed and started %s (unit: %s) on machine: %s", *serviceName, unitPath, machineId)
-	log.Printf("   systemctl status %s.service", *serviceName)
+func uninstallCmd(args []string) {
+	agentinstall.UninstallLinux(args, defaultServiceName)
 }
 
 func runCmd(args []string) {
@@ -131,258 +58,4 @@ func runCmd(args []string) {
 	}()
 
 	runAgent(*configPath, stopCh)
-}
-
-func uninstallCmd(args []string) {
-	fs := flag.NewFlagSet("uninstall", flag.ExitOnError)
-	serviceName := fs.String("service-name", defaultServiceName, "systemd service name")
-	unitDir := fs.String("unit-dir", defaultUnitPath, "systemd unit directory")
-	configPath := fs.String("config", defaultConfigPath, "path to config.json")
-	fs.Parse(args)
-	configPathExplicit := isFlagExplicitlySet(fs, "config")
-
-	mustBeRoot()
-
-	if !strings.HasPrefix(*unitDir, "/") {
-		log.Fatalf("--unit-dir must be an absolute path: %s", *unitDir)
-	}
-	if !strings.HasPrefix(*configPath, "/") {
-		log.Fatalf("--config must be an absolute path: %s", *configPath)
-	}
-
-	unitName := *serviceName + ".service"
-	unitPath := filepath.Join(*unitDir, unitName)
-	unitExists := false
-	binaryPath := ""
-	if _, err := os.Stat(unitPath); err == nil {
-		unitExists = true
-		path, err := binaryPathFromSystemdUnit(unitPath)
-		if err != nil {
-			log.Printf("failed to read binary path from unit file %s: %v", unitPath, err)
-		} else {
-			binaryPath = path
-		}
-		if !configPathExplicit {
-			path, err := configPathFromSystemdUnit(unitPath)
-			if err != nil {
-				log.Printf("failed to read config path from unit file %s: %v", unitPath, err)
-			} else {
-				*configPath = path
-			}
-		}
-	} else if !os.IsNotExist(err) {
-		log.Fatalf("failed to stat unit file %s: %v", unitPath, err)
-	}
-
-	if _, err := exec.LookPath("systemctl"); err == nil {
-		if unitExists {
-			log.Printf("Removing systemd unit %s", unitPath)
-
-			if err := runCmdLogged("systemctl", "stop", unitName); err != nil {
-				log.Printf("systemctl stop failed for %s: %v", unitName, err)
-			}
-			if err := runCmdLogged("systemctl", "disable", unitName); err != nil {
-				log.Printf("systemctl disable failed for %s: %v", unitName, err)
-			}
-			if err := os.Remove(unitPath); err != nil && !os.IsNotExist(err) {
-				log.Fatalf("failed to remove unit file %s: %v", unitPath, err)
-			}
-			if err := runCmdLogged("systemctl", "daemon-reload"); err != nil {
-				log.Fatalf("systemctl daemon-reload failed: %v", err)
-			}
-			if err := runCmdLogged("systemctl", "reset-failed", unitName); err != nil {
-				if !isUnitNotLoadedError(err) {
-					log.Printf("systemctl reset-failed failed for %s: %v", unitName, err)
-				}
-			}
-		} else {
-			log.Printf("No unit file found at %s; nothing to remove", unitPath)
-		}
-	} else {
-		if unitExists {
-			log.Printf("systemctl not found; removing unit file %s only", unitPath)
-			if err := os.Remove(unitPath); err != nil && !os.IsNotExist(err) {
-				log.Fatalf("failed to remove unit file %s: %v", unitPath, err)
-			}
-		} else {
-			log.Printf("systemctl not found and no unit file at %s", unitPath)
-		}
-	}
-
-	if err := os.Remove(*configPath); err != nil && !os.IsNotExist(err) {
-		log.Fatalf("failed to remove config file %s: %v", *configPath, err)
-	}
-	log.Printf("Removed config file %s", *configPath)
-
-	if binaryPath == "" {
-		exe, err := os.Executable()
-		if err == nil {
-			exe, err = filepath.EvalSymlinks(exe)
-			if err == nil {
-				binaryPath = exe
-			}
-		}
-	}
-	if binaryPath != "" {
-		if err := os.Remove(binaryPath); err != nil && !os.IsNotExist(err) {
-			log.Fatalf("failed to remove binary %s: %v", binaryPath, err)
-		}
-		log.Printf("Removed binary %s", binaryPath)
-	}
-
-	log.Printf("Uninstall completed for service %s", *serviceName)
-}
-
-// --- helpers ---
-
-func mustBeRoot() {
-	if os.Geteuid() != 0 {
-		log.Fatal("this command must be run as root (try: sudo ...)")
-	}
-}
-
-func renderSystemdUnit(exePath, configPath string) string {
-	// Root-running service, with moderate hardening.
-	// You can tighten further once you know all file paths the agent needs to write.
-	return fmt.Sprintf(`[Unit]
-Description=CertKit Agent
-After=network-online.target
-Wants=network-online.target
-
-[Service]
-Type=simple
-ExecStart=%s run --config %s
-Restart=always
-RestartSec=5
-
-# Hardening
-NoNewPrivileges=true
-PrivateTmp=true
-ProtectHome=true
-ProtectControlGroups=true
-ProtectKernelTunables=true
-ProtectKernelModules=true
-LockPersonality=true
-MemoryDenyWriteExecute=true
-RestrictRealtime=true
-RestrictSUIDSGID=true
-
-StateDirectory=certkit-agent
-LogsDirectory=certkit-agent
-
-[Install]
-WantedBy=multi-user.target
-`, shellEscape(exePath), shellEscape(configPath))
-}
-
-func shellEscape(s string) string {
-	// systemd unit files treat ExecStart as a command line; spaces matter.
-	// Easiest safe approach: wrap in quotes and escape embedded quotes/backslashes.
-	// This is conservative and works well for typical paths.
-	s = strings.ReplaceAll(s, `\`, `\\`)
-	s = strings.ReplaceAll(s, `"`, `\"`)
-	return `"` + s + `"`
-}
-
-func binaryPathFromSystemdUnit(unitPath string) (string, error) {
-	data, err := os.ReadFile(unitPath)
-	if err != nil {
-		return "", err
-	}
-
-	for _, line := range strings.Split(string(data), "\n") {
-		line = strings.TrimSpace(line)
-		if !strings.HasPrefix(line, "ExecStart=") {
-			continue
-		}
-
-		cmdLine := strings.TrimSpace(strings.TrimPrefix(line, "ExecStart="))
-		if cmdLine == "" {
-			return "", fmt.Errorf("empty ExecStart")
-		}
-
-		if strings.HasPrefix(cmdLine, `"`) {
-			end := strings.Index(cmdLine[1:], `"`)
-			if end < 0 {
-				return "", fmt.Errorf("invalid quoted ExecStart: %s", cmdLine)
-			}
-			return cmdLine[1 : 1+end], nil
-		}
-
-		fields := strings.Fields(cmdLine)
-		if len(fields) == 0 {
-			return "", fmt.Errorf("invalid ExecStart: %s", cmdLine)
-		}
-		return fields[0], nil
-	}
-
-	return "", fmt.Errorf("ExecStart not found")
-}
-
-func configPathFromSystemdUnit(unitPath string) (string, error) {
-	data, err := os.ReadFile(unitPath)
-	if err != nil {
-		return "", err
-	}
-
-	for _, line := range strings.Split(string(data), "\n") {
-		line = strings.TrimSpace(line)
-		if !strings.HasPrefix(line, "ExecStart=") {
-			continue
-		}
-
-		cmdLine := strings.TrimSpace(strings.TrimPrefix(line, "ExecStart="))
-		if cmdLine == "" {
-			return "", fmt.Errorf("empty ExecStart")
-		}
-
-		if strings.Contains(cmdLine, `--config "`) {
-			parts := strings.SplitN(cmdLine, `--config "`, 2)
-			if len(parts) == 2 {
-				end := strings.Index(parts[1], `"`)
-				if end >= 0 {
-					return parts[1][:end], nil
-				}
-			}
-		}
-
-		fields := strings.Fields(cmdLine)
-		for i := 0; i < len(fields); i++ {
-			if fields[i] == "--config" && i+1 < len(fields) {
-				return strings.Trim(fields[i+1], `"`), nil
-			}
-			if strings.HasPrefix(fields[i], "--config=") {
-				return strings.Trim(strings.TrimPrefix(fields[i], "--config="), `"`), nil
-			}
-		}
-		return "", fmt.Errorf("--config not found in ExecStart")
-	}
-
-	return "", fmt.Errorf("ExecStart not found")
-}
-
-// If you want to give a nicer error when systemctl isn't present.
-func isCmdNotFound(err error) bool {
-	var ee *exec.Error
-	if errors.As(err, &ee) {
-		return true
-	}
-	return false
-}
-
-func isUnitNotLoadedError(err error) bool {
-	if err == nil {
-		return false
-	}
-	return strings.Contains(strings.ToLower(err.Error()), "not loaded")
-}
-
-func isFlagExplicitlySet(fs *flag.FlagSet, name string) bool {
-	wasSet := false
-	fs.Visit(func(f *flag.Flag) {
-		if f.Name == name {
-			wasSet = true
-		}
-	})
-	return wasSet
 }
